Fix floor returning float in python2
Traceback (most recent call last):
  File "./ccs.py", line 261, in <module>
    CcsInteractive(players).cmdloop()
  File "/usr/lib/python2.7/cmd.py", line 142, in cmdloop
    stop = self.onecmd(line)
  File "/usr/lib/python2.7/cmd.py", line 221, in onecmd
    return func(arg)
  File "./ccs.py", line 181, in do_state
    pretty_print_games(self.players)
  File "./ccs.py", line 147, in pretty_print_games
    n=str(i + 1).rjust(len_players_col)
TypeError: integer argument expected, got float

+++ ccs.py
@@ -139,7 +139,7 @@ def pretty_print_games(players):
     #             players     +  1st column  + score  + ties  + sep
     print('-' * (nplayers * 8 +  18          + 6      + 10    + 5))
 
-    len_players_col=math.floor(math.log10(len(players))) + 1
+    len_players_col=int(math.floor(math.log10(len(players))) + 1)
     # table
     i = 0
     for p in players:
